CQRS 기본 적용 & Port 방식 도입
