+++ delivery/src/main/java/com/hubEleven/delivery/application/port/DeliveryQueryPort.java
@@ -1,0 +1,8 @@
+package com.hubEleven.delivery.application.port;
+
+import com.hubEleven.delivery.domain.Delivery;
+import java.util.UUID;
+
+public interface DeliveryQueryPort {
+    Delivery delivery(UUID id);
+}
